Moved templated function implementations from the data_reader_HDF5.hpp into a data_reader_HDF5_impl.hpp.
