+++ include/lbann/data_readers/CMakeLists.txt
@@ -6,6 +6,7 @@ set_full_path(THIS_DIR_HEADERS
   data_reader_csv.hpp
   data_reader_image.hpp
   data_reader_HDF5.hpp
+  data_reader_HDF5_impl.hpp
   data_reader_merge_features.hpp
   data_reader_merge_samples.hpp
   data_reader_mnist.hpp
