+++ src/proto/proto_common.cpp
@@ -34,7 +34,7 @@
 #include "lbann/utils/file_utils.hpp"
 #include "lbann/utils/argument_parser.hpp"
 
-#include "lbann/data_readers/data_reader_HDF5.hpp"
+#include "lbann/data_readers/data_reader_HDF5_impl.hpp"
 
 #include <lbann.pb.h>
 #include <reader.pb.h>
